Align string generate method with generate_json_string
